+++ src/src/devtools.ts
@@ -1,6 +1,6 @@
 chrome.devtools.panels.create(
     'http helper',
-    'vue_dst/icons/icon16.png',
-    'vue_dst/panel.html',
+    'dist/icons/icon16.png',
+    'dist/panel.html',
     () => console.log('[http helper] Panel created')
 )